Include GAP Id for groups
There are multiple groups with the same semi-direct product description, so some groups are lost in the drop down list.

This fixes the problem by also including the GAP Id for the group.

Future work: GAP already has a method for producing the subgroup lattice. Might be faster to wrap this instead.